Euklides i symbol Newtona

+++ silnia_reku.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 
-int silnia(int x){
+long long silnia(int x){
     if(x==0){
         return 1;
     }else if(x<0){
